+++ src/types/storeDataTypes.ts
@@ -11,14 +11,3 @@ export interface SignUpStoreStateType {
   phoneNumber: string;
   gender: string;
 }
-
-export interface SignUpStepType {
-  stepId: number;
-  isEnable: boolean;
-  item: React.FC<{
-    step: number;
-    handleChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-    errorMessages: Partial<SignUpStoreStateType>;
-    inputValues: SignUpStoreStateType;
-  }>;
-}
